test: add paddle specific tests
