+++ tests/unit/document/paddle/test_ndarray.py
@@ -1,8 +1,6 @@
 import numpy as np
+import paddle
 import pytest
-import tensorflow as tf
-import torch
-from scipy.sparse import csr_matrix, coo_matrix, bsr_matrix, csc_matrix
 
 from docarray import Document
 
@@ -11,15 +9,7 @@ def get_ndarrays():
     a = np.random.random([10, 3])
     a[a > 0.5] = 0
     return [
-        (a, False),
-        (torch.tensor(a), False),
-        (tf.constant(a), False),
-        (torch.tensor(a).to_sparse(), True),
-        (tf.sparse.from_dense(a), True),
-        (csr_matrix(a), True),
-        (bsr_matrix(a), True),
-        (coo_matrix(a), True),
-        (csc_matrix(a), True),
+        (paddle.to_tensor(a), False),
     ]
